finished e2e nft test script
Closes #571
